All robustness test engine interface definitions moved into the test root directory.
