+++ tests/robustness/comparer.go
@@ -1,6 +1,4 @@
-// Package checker defines the framework for creating and restoring snapshots
-// with a data integrity check
-package checker
+package robustness
 
 import (
 	"context"
