+++ tests/tools/kopiarunner/kopia_snapshotter.go
@@ -19,10 +19,10 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/kopia/kopia/internal/retry"
-	"github.com/kopia/kopia/tests/robustness/snap"
+	"github.com/kopia/kopia/tests/robustness"
 )
 
-var _ snap.Snapshotter = &KopiaSnapshotter{}
+var _ robustness.Snapshotter = &KopiaSnapshotter{}
 
 const (
 	contentCacheSizeMBFlag  = "--content-cache-size-mb"
